git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@9501 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/neighbor.cpp
@@ -1557,7 +1557,7 @@ void Neighbor::setup_bins()
 
   // memory for bin ptrs
 
-  bigint bbin = mbinx*mbiny*mbinz;
+  bigint bbin = ((bigint) mbinx) * ((bigint) mbiny) * ((bigint) mbinz);
   if (bbin > MAXSMALLINT) error->one(FLERR,"Too many neighbor bins");
   mbins = bbin;
   if (mbins > maxhead) {
